modify: gtest/

+++ gtest/inst.sh
@@ -6,5 +6,11 @@ pacman -S --needed --noconfirm \
 	$MINGW_PACKAGE_PREFIX-gtest \
 	lcov
 cp -p ../strconv.h $MINGW_PREFIX/include/
-rm -rf nlohmann
-cp -rp $MINGW_PREFIX/include/nlohmann .
+
+sts=0
+diff -r $MINGW_PREFIX/include/nlohmann ./nlohmann || sts=$?
+echo sts=$sts
+if [ $sts -ne 0 ]; then
+  rm -rf nlohmann
+  cp -rp $MINGW_PREFIX/include/nlohmann .
+fi
